Switch link-visited-color and link-color
Hi Daniel! Excellent fork, I'm enjoying it a lot!

It is my understanding that from an accessibility point of view, **visited links** should melt a bit with its background as a visual indicator to the user that he's already been there.

Following that logic, the **exact opposite** is expected for dark themes, therefore visited links should be lighter that regular ones.