Do not strip `indent` attribute
HN is now using an attribute `indent` to identify how deeply nested a
comment is. Since it will repeatedly check the next element and compare
indent levels to see if it should additionally hide that element, we
cannot strip this without breaking folding.

When stripping out `indent`, only the targeted comment is hidden, none
of the sub-elements.